% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denseburn-package.R
\docType{package}
\name{denseburn-package}
\alias{denseburn}
\alias{denseburn-package}
\title{denseburn: Sparse Polygon Rasterization with Exact Coverage Fractions}
\description{
Computes exact coverage fractions for polygon-grid intersections and returns results in a sparse two-table format: run-length encoded interior cells and individually weighted boundary cells. Uses the exactextract algorithm by Daniel Baston, vendored from the 'exactextractr' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/denseburn}
  \item Report bugs at \url{https://github.com/hypertidy/denseburn/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Daniel Baston (exactextract C++ algorithm (vendored)) [copyright holder]
  \item ISciences, LLC (exactextract C++ algorithm (vendored)) [copyright holder]
}

}
\keyword{internal}
